requirements.txt
selenium==4.15.2
minio==7.2.0
beautifulsoup4==4.12.2
requests==2.31.0
psycopg2-binary
python-dotenv
sqlalchemy